Seperate child components by role
