+++ src/App.vue
@@ -11,12 +11,16 @@ export default {
 </script>
 
 <style>
+  html, body {
+    height:100%;
+  }
 #app {
   font-family: 'Avenir', Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: #2c3e50;
+  height: 100%;
   margin-top: 60px;
 }
 </style>
